+++ settings.py
@@ -11,22 +11,6 @@ REMOTE_SOURCES = {
     'patients':{
         'url':'https://www.harp.lg.jp/opendata/dataset/1369/resource/2828/patients.csv',
         'type':'csv'
-    },
-    'patients_summary':{
-        'url':'https://www.harp.lg.jp/opendata/dataset/1369/resource/2829/patients_summary.csv',
-        'type':'csv'
-    },
-    'discharges_summary':{
-        'url':'https://www.harp.lg.jp/opendata/dataset/1369/resource/2833/discharges_summary.csv',
-        'type':'csv'
-    },
-    'current_patients':{
-        'url':'https://www.harp.lg.jp/opendata/dataset/1369/resource/2832/current_patients.csv',
-        'type':'csv'
-    },
-    'inspections':{
-        'url':'https://www.harp.lg.jp/opendata/dataset/1369/resource/2837/inspections.csv',
-        'type':'csv'
     }
 }
